requirements.txt
matplotlib==3.7.2
networkx==3.1
numpy==1.23.5
PennyLane==0.31.1
qiskit==1.2.0
scipy==1.16.1
tensorcircuit==0.10.0
tensorflow_intel==2.11.0
torch==2.0.1
